backend/requirements.txt
# Backend requirements
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.4.2
